early check in added code to delete tmp files and allow caller to provide list of sample names

+++ bin/mineSalmonLogs.sh
@@ -36,18 +36,10 @@ set -x # turn debug on
 # set + x # turn debug off
 
 listOfSampleNames=$2
-echo AEDWIP $listOfSampleNames AEDWIP
 
-if [[ ! -f $listOfSampleNames ]]
-then
-    echo listOfSampleNames does not exist
-else
-    echo we got a file
-fi
-
-exit
 
 source createTmpFile.sh
+listOfTmpFiles = ""
 
 # kl=/private/groups/kimlab
 # krasDir="${kl}/kras.ipsc/data"
@@ -67,18 +59,31 @@ source createTmpFile.sh
 # TODO AEDWIP pick mapping rate our of aux_info/meta_info.json instead of log file
 grepOutTmp=`createTmpFile`
 add_on_exit rm  $grepOutTmp
+listOfTmpFiles = "$listOfTmpFiles $grepOutTmp"
+
 grep -i "mapping rate" ${salmonLogs} > $grepOutTmp
 
-
+#
+# create sample name column data
+#
 sampleNameTmp=`createTmpFile`
-printf "sampleName\n" > $sampleNameTmp
-add_on_exit rm $sampleNameTmp
-cut -d / $grepOutTmp -f 5,6,7,8,9 >> $sampleNameTmp
+listOfTmpFiles = "$listOfTmpFiles $sampleNameTmp"    
 
-salmonOutTmp=`createTmpFile`
-add_on_exit rm $salmonOutTmp
-printf "salmonOut\n" > $salmonOutTmp
-cut -d / $grepOutTmp -f 10 >> $salmonOutTmp
+add_on_exit rm $sampleNameTmp
+printf "sampleName\n" > $sampleNameTmp
+    
+if [[ ! -f $listOfSampleNames ]]
+then
+    cut -d / $grepOutTmp -f 5,6,7,8,9 >> $sampleNameTmp
+
+    salmonOutTmp=`createTmpFile`
+    add_on_exit rm $salmonOutTmp
+    printf "salmonOut\n" > $salmonOutTmp
+    cut -d / $grepOutTmp -f 10 >> $salmonOutTmp
+else
+    cat $listOfSampleName >> $sampleNameTmp
+fi
+
 
 # https://stackoverflow.com/a/41996668/4586180
 #echo $foo | sed -n -e 's/^.*\(\(Mapping\).*\)/\1/p'
@@ -87,6 +92,7 @@ cut -d / $grepOutTmp -f 10 >> $salmonOutTmp
 # cut is a hack to pick out the numeric value
 # use tr to get rid of percent
 mappingRateTmp=`createTmpFile`
+listOfTmpFiles = "$listOfTmpFiles $mappingRateTmp"
 add_on_exit rm $mappingRateTmp
 printf "mappingRate\n" > $mappingRateTmp
 sed -n -e 's/^.*\(\(Mapping\).*\)/\1/p' $grepOutTmp  | cut -d = -f2 | tr "%" " " >> $mappingRateTmp
@@ -96,14 +102,17 @@ sed -n -e 's/^.*\(\(Mapping\).*\)/\1/p' $grepOutTmp  | cut -d = -f2 | tr "%" " "
 # we use a for loop so taht we find the correct cmd_info.json file
 #
 indexTmp=`createTmpFile`
+listOfTmpFiles = "$listOfTmpFiles $indexTmp"
 add_on_exit rm $indexTmp
 printf "index\n" >> $indexTmp
 
 mates1Tmp=`createTmpFile`
+listOfTmpFiles = "$listOfTmpFiles $mates1Tmp"
 add_on_exit rm $mates1Tmp
 printf "mate1\n" >> $mates1Tmp
 
 mates2Tmp=`createTmpFile`
+listOfTmpFiles = "$listOfTmpFiles $mates2Tmp"
 add_on_exit rm $mates2Tmp
 printf "mate2\n" >> $mates2Tmp
 
@@ -144,6 +153,7 @@ done
 # count the number of unmapped reads
 #
 readCountTmp=`createTmpFile`
+listOfTmpFiles = "$listOfTmpFiles $readCountTmp"
 add_on_exit rm $readCountTmp
 masterCountReads.sh $salmonLogs > $readCountTmp
 
@@ -151,10 +161,17 @@ masterCountReads.sh $salmonLogs > $readCountTmp
 # create output file
 #
 pasteTmp=`createTmpFile`
+listOfTmpFiles = "$listOfTmpFiles $pasteTmp"
 add_on_exit rm $pasteTmp
 
 paste $sampleNameTmp $mappingRateTmp $salmonOutTmp $readCountTmp $indexTmp $mates1Tmp $mates2Tmp > $pasteTmp
 cat $pasteTmp
+printf "\n"
 
 
+# clean up tmp files
+for t in $listOfTmpFiles;
+do
+    'rm' $i
+done
